VBV-1235 - Sending placement zone queries that exceeds size limit
The problem was that after the UI constructs the odata request the url
is too long and we get an error from netty.
- Created a new stateless services that accepts post request with body
the url and then return the result. Once the request reach the xenon
node the size of the url is not a problem.
Tested also in cluster and it works as expected the odata query is
executed always on the same node.

This is the first part of the fix. The second part is to change to UI to
call the new service instead of directly sending odata request.

Change-Id: Id13e5e29bdd69f22d445195934a0cb767fee2be6
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/10882
Reviewed-by: Stanislav Hadjiiski <hadjiiskis@vmware.com>
Compute-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>