VBV-1243 Resources are not cleaned up when application deployment fails
Container deployment in this case on container start. Thus the container
is deployed on the docker host, but is not running. We don't update the
container id between the deploy and start operations, so we can't delete
the container if we fail on the latter

Change-Id: I01cf7810e87efdb9113b592f25cff00f793c71cb
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/10964
Compute-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Stanislav Hadjiiski <hadjiiskis@vmware.com>