Create general delete confirmation component. Use it for project deletion
Implemented a delete-confirmation component - a general modal dialog with
customizable title and description with "cancel" and "remove" options. The
dialogs fires "onDelete" and "onCancel" events corresponding to the user
action. It is also possible to provide an "alert" (e.g. error message for
failed operation). Once the "remove" button is clicked, all controls are
disabled: buttons are grayed out, there is a spinner in the "remove"
button, the close button disappears and the dialog cannot be canceled by
clicking outside of it. The dialog will be closed on removal success (by
setting the "visible" property to false). Alternatively, dialog controls
will be again enabled on failure (after an "alert" is set).

Also, removed the project-delete component and put the new component on
the line.

TODO currently the loading spinner in the "remove" button is done
manually. However, in Clarity 0.9.2+ there is a "clrLoading" angular
property that can be set to true or false to display/hide the same
spinner out of the box. When clarity gets updated, the code can be
refined.

Change-Id: I188828b031b4da93c7909c702ef3770dc9a3058f
VIDEO: http://url.vmware.com/gt63
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/11170
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Compute-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Iveta Ilieva <iilieva@vmware.com>