Assing/Unassing groups to project roles.
Reworked ProjectState and ProjectService, to hold list of groups,
for admins and members.

When assigning or unassigning users/groups to project it will return
status code 200 in case if it was successful instead of 304.

When attempt to assign users that does not exist in the AD or local users,
it will return 400 instead of 500.

Introduced PrincipalNotFoundException to be used as common exception when
searching for not existing principal in either PSC or local users.

Change-Id: Iee172436a83701c86d8255e15d73ebd85cd2a77d
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/11769
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Stanislav Hadjiiski <hadjiiskis@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>