[VBV-1439][UI] Link the Project selector to a custom header with the selected project
On each request to the back-end, the currently selected project is set
in the x-project header. It is up to the back-end to determine
whether this information is needed for the requested resource or
operation.

Removed the option to chose project when provisioning a container or
application.

Additionally, added the /requests and /request-status endpoints to the
npm proxy configuration so resources can now be provisioned while
developing the UI.

Change-Id: I137e9ffff5ecca9736e86b9ba50c9c52173a8088
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/12624
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Antonio Filipov <afilipov@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>