Refactored CIDR allocation service to use subnet link
This changeset basically reverts a previous change that was made to use subnet id instead of link as the subnet identifier for CIDR allocation. The problem with that approach was that the subnet id was reset by the photon adapter to the actual/real subnet id after successfully provisioning the subnet, preventing the CIDR to be de-allocated when the subnet was destroyed (as no match was found in the CIDR allocation map).
In addition to reverting to using the subnet link, this fix modifies the compute network provisioning task to generate the subnet link prior to allocating the CIDR.

Change-Id: I31790b6dd5fc5023a84298e9dd7d55f714e31ad8
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/13353
Compute-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Valentina Reutova <vreutova@vmware.com>