Fix MigrationIT
The problem was that the updateTimeMicros was compared and this field is
taken from the source host during the migration. So it is not possible
to rely on it in order to determine when the document was last updated
on the target env.
- Removed the check for the updateTimeMicros.

Change-Id: I23424c0f58c0284047f3db7239b7897c1108a0e3
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/13428
Reviewed-by: Zahari Ivanov <zaharii@vmware.com>
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>