When adding known official docker registry do not append default port.
The change is in both the back end and UI.

Added additional check when setting the UI title to either
Admiral or VIC, if it's not embedded.

Added additional check in main-resources-component, if the
current user is project viewer and there is no harbor configured
to not show the Library in left navigation bar, because in this case
none of Library submenus should be shown.

Change-Id: I718fea64ecef1dcdf6560189a0f58f373968fe5f
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/15771
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>