Fix for tenantLuinks of the resource pools when upgrading vRA
The problem was that after upgrade the tenant links of the resource
pools were not updated based on the placement and because of this the
clusters were not visible in the UI. The problem appear when the pool
has already tenant links. In this case we used to set them for the
resource pool but not for the EPZ. In this case the EPZ service
overrides the tenant links for the resource pool.
- Set the EPZConfig tenant links to null before sending the request
- Added unit test for the case.

Change-Id: Iee6fe486569c81bccb50f9a294b909c5ad4a7844
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/15943
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>