fix github issue 183 - Problems upgrading appliance from 1.1.1 to 1.2
The problem was that after upgrade it was not possible to load the
projects. The reason was that during the migration of the data we send
post for the default project to the new admiral instance. On the project
service we have idempotent post and the default project on the new
system was overwritten. This is a problem because we have new fields in
the project states that are required in order to have the project
working properly.
- Added check in order to avoid update of the default project during
upgrade.
- Added unit test for the case
- Tested locally that after upgrade the environment is working as
expected.

Change-Id: Idf5505a70a4741497180963bd38d7227e11814b7
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/16172
Closures-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>