Try to fix upgrade job
The job is failing sometimes with Failure when calling migration task.
It could be that when trying to do the migration the nodes are not
started yet.
- Increased the time to wait before doing the migration. Usually we use
the health check service however the job validates upgrade from a
version where the health check was not available.
- Added the original exception. Like this it will be more visible in
case of failure what is the problem.

Change-Id: I939935d1276e7deada8974c35a52a3ab0747f0ec
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/16434
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Stanislav Hadjiiski <hadjiiskis@vmware.com>