Fix migration failing with Service not found:/management/migration-tasks
The issue happen in the last run of cat. The problem is that before
starting the migration we check the statuses of the services that are
going to be migrated but we did not check the status of the migration
task service. In the logs of the failed cat it is visible that migration
was attempted and after the failure the migration service was started.
- Added the migration task to the list of the services to check.

Change-Id: If0838fec63814d7b9e552d0fc03f97058c60b2fc
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/17222
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>