Fix overlap issue with long log-control labels in some localizations
In some localizations, the log control labels are too large so when the
screen size is smaller, there is not enough space for all the content
and part of it is wrapped. That part appeared over the actual logs. To
fix this, the labels are now shrinking when the window is resized. The
overflown text is hidden in ellipsis. Also, hovering the mouse over
the label reveals the full text.

SCREENSHOT: http://url/tt23

Change-Id: I936987e7c99a6121c09ae490e336752a20f60ca2
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/17294
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>