VBV-1587 - Configuration properties read from file do not update ConfigurationStates
As described in the issue, the configuration states should be updated
when the corresponding configuration files are updated, and they should
be the ultimate source of truth. If some property defined in a
configuration file is changed via REST API and the change is supposed
to be kept after restarts, the proper configuration file needs to be
updated.

Also, marked the ConfigurationUtil main methods as deprecated to try to
avoid using them since they rely on a memory cache of the configuration
properties instead of the actual configuration property values (states).
This can bring some inconsistencies or unexpected behavior as described
in the Javadoc.

Change-Id: Ic0386230d5e86a424ac816066a4a6f84bafa8bc7
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/17708
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>