Introduce public address for hosts
As mentioned in Gtihub issue #12 Container details - Address - Ports,
published container ports are not always accessible on the same address
that exposes the docker API. This changge introduces the option to set
the public address of a host (the one that exposes the published ports)
in the UI. The address is stored as a custom property of the
ComputeState. If the property is missing, the public address is derrived
from the ComputeState's address (the one that is used for the Docker
API).

VIDEO: http://url/z4m3

TODO: There is currently no option to change the public address for
clusters that permit a single container host in them (e.g. VCH
clusters). In the UI, a public address field is added to the edit VCH
cluster dialog, however, some backend changes are needed to propagate
that change to the underlying host. For the case of docker clusters
(multiple hosts allowed) everything is there. The public address can be
ammnedned by editing the host from the resources tab of the cluster
details view.

Change-Id: I32c8e55b6891fedeb5bd7ba7ae0da6d6c0f60eee
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/18294
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>