Upgrade: fix a problem when during upgrade xenon node tries to join in cluster
The problem is that during upgrade when one node start the upgrade
procedure it could happen that the second node can try to start in
cluster. This results in a problem because the quorum is changed in the
middle of the upgrade/migration and in this case the upgrade fails.
We already stop the health check service in order to prevent that
however there is some small period of time when this could happen.
The solution is to make the health check service return 500 during the
upgrade. and use a different method to get the health before starting
the upgrade
- Added property. When specified the health check will return 500 for
get
- Added handlePost in order to use it in case of upgrade

Next step:
- Change the way we start in the appliance.

Change-Id: Ie9051ac481a1f9235ecb83ad0538d4c259435510
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/19389
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>