VBV-1708: Wrong /or VERY STRANGE/ placement of containers
The issue happens when there are several clusters with same priority. In this
case, the same cluster is always chosen (the first one returned by the query,
which is alwas the same). This is breaking performance tests, because in that
test the hosts are added in different clusters (n clusters with one host each
one), so all provisioned containers go to the same host until the host dies.

The fix is to shuffle all the placements with maximum priority, so in method
selectReservation we choose a random placement between the ones with max
priority.

Change-Id: I5629f9e5246bb5aa6ad8f4f68c0e181b7d7c702d
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/19639
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>