Fixed issue #79 -  Admiral fails deploying containers with non-CIDR networks
The problem originally described in the issue occurred when a new VCH
host was created and the creation command includes the container network
definition with the --container-network-ip-range in a non-CIDR format
(as described in https://vmware.github.io/vic-product/assets/files/html/1.2/vic_vsphere_admin/container_networks.html).

In such scenario, the VCH API reported the IP range in a format that
was not supported by Admiral (since the standard Docker API doesn't
support it either). Because of that the network was discovered by
Admiral but it was left in an "Unknown" state and unusable by the
applications.

On the other hand, the creation of networks once the VCH is created,
through the Docker CLI (against VCH) or the Admiral UI (also against
VCH), behave in the same way and they fail when trying to use a
non-CIDR annotation for the IP range (as it's expected from the Docker
API).

With these changes now the non-CIDR format it recognized properly in
the scenario described above, and applications can make use of the
discovered network (which also appears as "Connected").

No changes are needed from the UI point of view since, again, when
creating networks on existing hosts, both VCH and Docker APIs behave
in the same way.

Added some new tests.

Change-Id: I0d13f3d202efcd566da3f748f6002709d6f2f169
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/20401
Reviewed-by: Stanislav Hadjiiski <hadjiiskis@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/26541
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>