Use new serviceClient for each host ping.
There is a bug 1983364 which happens when same credentials are used
for two hosts. Because the credentials are cached it is possible to add
a second host without specefying the credentials.
Also if a host is added normaly (with credentials), deleted and then
added again  for the second adding is possible to omit the credentials.

Change-Id: I248378a6261cf148031d1ee975df8de65f8f6aca
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/21100
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>