VBV-1706 - Wrong (and random) host memory values are displayed in UI
The problem was that on provisioning/removing a container the max/total
memory for the placement were updated using the logic : if add operation
add the memory of the host to the total in case of remove: decrease the
total memory. Because of this during provisioning we were observing
double, triple and so on amount of memory for the cluster.
- Added a flag in order to make the difference when the the operation is
over container
- With the flag in case of container provisioning/removal we will not
modify incorrectly the total memory of the placement.

Change-Id: I84b9ae8c89c9e19102ce8a358cc00da5821c8558
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/21191
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>