Prevent volumes created in one project to be seen in other projects
If there is a host with volumes added in a project and if that host is
added to another project, the volumes are visible in the second project.
Containers and networks are not visible in such scenario and we need
consistent behaviour.
Remark: In this context only the name of the volume is enough to distinguish
the volumes, because for created by admiral containers the volume service id
is used.
https://jira-hzn.eng.vmware.com/browse/VBV-1735

Change-Id: I688fc6cffe3311326953f2e62ce8655d4f1b7d5c
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/21877
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>