VBV-1742 - [UI] When creating a cluster, the memory bar is full and label is 8 of 8EB instead of N/A
The problem was that when a host memory was to available Long.MAX_VALUE
was set. In order to display N/A the value should be 0. One the real
memory is available the correct value is set for the cluster.

Change-Id: I572c515af8525f436c347094a46d995135a5cb47
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/21691
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>