Fix: scale out container shows two request
When we scale out container in the request bar in the ui
on the left we see that two request are started. The
reason is that in the ClusteringTaskService we miss to
give the requestBroker requestTrackerLink. When this link
is missing the requestBroker create new requestTracke.

Change-Id: Ica1cc77a6d11d3a09021bf16d9d5ac040102972f
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/22555
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>