VBV-1710 - Session not terminated on the server side
The session is indeed terminated on the Admiral side when doing a logout
but the Xenon auth caching mechanism is expecting to be valid until it
expires. Because of that Xenon may still "accept" sessions after doing
logout for them. To workaround that in VIC mode, I've added a cache of
invalidated sessions so that, when necessary, a check is performed
against that cache and the incoming request is unauthenticated if
needed.

Change-Id: I15041f238aaf3f839092d75db9771f95dd68aee3
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/22554
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>