VBV-1764 - [VIC] Voting app fails only when the host is VCH
The original problem is that the provisioning in VCH is slow, and while
the containers, networks and volumes of the application can be created
apparently OK, internally in VCH the connection of a container to a
network can fail if those entities are not fully initialized.

With that, we didn't have in place a retry logic for the operations of
(on create) connecting a container to a network and starting the
container after that.

The fix includes those retries and some small renaming.

After the fix the voting app is successfully deployed on VCH (the 1st
time usually slowly because images are missing usually, the next times
much faster).

Change-Id: Iffcd1441096f3dd9cddc55fb1e122f71dd4d12ed
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/23316
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>