Potentially related to Admiral went to inaccessible state while removing VCHs sequentially - High CPUPerc #240
Added a couple of enhancements to the container descriptions:
- When doing the handle periodic maintenance: If after getting the
container description we see that it has no health config settings,
then we don't need to query for its associated container states because
we won't perform any health check at the end. In this way we'll save,
and I'd say that most of the times, a Xenon direct query for nothing on
every handle maintenance for each container description.
- When deleting a container (state): Delete its associated container
description if no other container states are associated to it as well.
In this way we'll avoid "zombie" container descriptions doing and
having scheduled handle maintenances for nothing.

This is an attempt to reduce the number of unnecessary container
descriptions that Admiral keeps and that have associated always
periodic maintenance tasks.

Change-Id: Ia537ba7d1150d81d98da74db7c8ffa77a91ec346
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/23842
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Antonio Filipov <afilipov@vmware.com>