DeploymentProfileConfig test property should be handled consistently
Many tests are modifying the test property in the DeploymentProfileConfig
singleton in an inconsistent way, leading to side effects running tests
(depending on the order tests are run, such property can be true or false for
the same test).

This fix sets the test property to true in BaseTestCase
@before, and resets to false in the @after method. Any test that inherits
BaseTestCase can change the property to false if required by the test, and it
is not mandatory to revert it back. Tests not inheriting BaseTestCase should
expect test property to be false. They can change it to true if needed, and in
this case they must set it back to false in an @after method.

Change-Id: Ibdba0b2188f85b06064e396b4ce2265c592336aa
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/24323
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>