Remove ClusterMonitoringService in order to work with 3 nodes
The service is not needed anymore because the new recommendation for
cluster for admiral is to use 1 or 3 nodes. Until now this service was
automatically decreasing the quorum to 1 in case there were no other
nodes in the node group. However this is not recommended by xenon and
the value was questionable.
Possible problems when using this service:
- In case we have cluster of 2 nodes in vRA and at some point xenon is
not available the quorum will be set to 1. However if the node continue
working and at some point the second node appear the qorum will be 2
again and in this case the nodes will not be able to agree about the
existing data.
- In case one xenon node is not available the quorum was set to 1
however because of the "sticky session" in vRA cluster the containers
tab will still not load in 50% of the cases.

Note: The start/stop script in vRA will continue working as before by
manipulating the quorum in case of 2 nodes cluster.

- Removed the ClusterMonitoringService

Change-Id: Ia02f0107054b9ae3d281138e7db0b57231bdaeec
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/24562
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>