[VBV-1814] Use public host address for healthcheck
Before this commit, if HTTP/TCP healthcheck was configured for a
container that exposed no ports, a wrong IP address would have been used
for the execution of the health check requests. That address was
retrieved from the "IPAddress" property of the docker inspect for that
container. However, this property is deprecated in docker since the
introduction of container networks because since then a container could
have multiple addresses (one in each network). On Docker hosts, this
address still resolves to the IP address on the bridge network, however,
it is in a private range (e.g. 127.1.0.3) so it cannot be reached from
Admiral. On VCHs, this property is empty. This was causing some error
messages for invalid healthcheck configuration to appear in the logs and
due to a code bug was also preventing the healthcheck from ever completing.

To address this, instead of the IP address property of the container,
the public address of the host will be used. If the public address
feature is disabled, then the host API address (the default address)
will be used. The public address of the host will be also used in the
case when the container has at least one published port (previously, the
host API address was used regardless of the public address feature
toggle).

Change-Id: I0aac3d7aa5cedabee84a9d09f111ee44467487f1
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/24665
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>