Introduce RetriableTask class
Related to issues #233 and #237

Currently, most requests to the Docker API are retried immediately on
failure. This was causing problems with VCH hosts which experienced
issues with stacks of multiple retry operation. Also, as a general rule
it is a good practice to have the retries sent after a small delay
(exponentially growing with the number of retries).

This change introduces the RetriableTask that automatically retries a
given task a configurable number of retries. The task is supposed to be
build with the RetriableTaskBuilder which provides a unified API for
configuring and executing retriable tasks.

Change-Id: If267c689855352027fed4786350025e98bb6b533
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/25698
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>