Retry container start commands with delay for newly created containers
Related to issue #237

Currently, we make difference between starting an old container and starting
a newly created one. The main difference was the addition of the
ExposedPorts and HostConfig.PortBindings properties in the JSON body of
the request to the Docker API. However, according to Docker
documentation (https://docs.docker.com/engine/api/v1.24/) these
properties need to be set for the Create Container command (and not for
the Start Container one).

Since we do set the aforementioned properties for the Container Create
command, and since this was the only difference between starting an old
and a new container, in this commit the duplicated functionality is
removed and the existing code for Container Start is reused.

As a side effect (and main goal of this commit), starting newly created
containers now makes use of the RetriableTask API and thus failed
requests are retried with the default delay (5, 10, 30 seconds).

Change-Id: I97e6be061483cccf02e55341abae8cadd38ecae5
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/25781
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>