Limit retries to a single retry for DockerAdapterService tests
Related to issues #233, #237, VBV-1831

With recent commits, retries for container operations submitted from
DockerAdapterService to the Docker API are now waiting for a certain
delay. This artificially increases the execution time for
DockerAdapterServiceTest with ~45 seconds for each test that involves a
retry logic.

In order to reduce the execution time for the test, the number of
retries is limited to 1 for the scope of the test. This does not fully
eliminate the artificial delay, but reduces it to ~3 seconds per test
case that involves retry logic. When the retry delays are made
configurable, the test can be tweaked again to reduce the retry delays
instead of the number of retries.

Change-Id: I3bf2bd5d0b8a2735829c1995283e3cf7f2bc0b44
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/25816
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>