Improve error messages in the Event log in case of RetriableException
With the new RetriableTask API, some calls to the Docker API are
retried. In case of failure, the original exception is stored as the
cause of a newly created RetriableException. The message of this
exception is not descriptive enough, e.g. it states that the maximum
number of retries has been exceeded, however it gives no clue about the
actual failure.

This commit gets message of the original exception and appends it to the
RetriableException's message. This makes it possible to see that the
exception was retried and also to know what was the failure in the first
place.

Picture: http://url/pjzf

Additionally:
- added test cases to increase test coverage
- now the retriable tasks will not be retried if the reason for failure
  was an "already executed" failure (i.e. violation of the single
  execution policy for the task)

Change-Id: I381cfdd91fe2aab3fe5644ed7197632812d8b229
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/26477
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>