Provide error details in case of unexpected error for container host operations
Related to #253

Currently, if an unexpected error occurs for an adapter request sent
from the ContainerHostService, the details for the actual error will be
masked (an empty string will be displayed both in the UI and in the
logs). This makes it harder to track causes of issues. This has been
reported by the community and is preventing us to track the reason for a
failed creation of a cluster (the ping fails with an unexpected error).

This commit introduces an "unexpected exception" message that will
contain the original error message. Also, the original error will be
available in the logs.

Change-Id: Ie2496c76f35e371886b7dd9654d556a88d156647
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/27043
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Stanislav Hadjiiski <hadjiiskis@vmware.com>