Allow consistent provisioning of a template with external network
Scenario: Create two clusters. Create a template with an external
network and a user defined network, have three containers - one
connected to both networks, one to the exernal only and one to the user
defined.

Note that if no KV store is configured, the whole app should be deployed
on the sam host - the one with the external network. The containers
though are not linked and the host for each will be determined
separately.

Add a dependency between containers using the same user-defined network
and update the host of that network once the first container is
provisoned.

Change-Id: I4cc37918dbe1b2ddf10a299f10f6899f6c5f2cd8
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/27794
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>