VBV-1860 - [Upgrade][HA] Application is not displayed after upgrade 7.3 -> 7.4: missing tenant link
The problem was that in 7.2/7.3 when there is a failed provisioning from
the catalog it is possible to have an empty application in admiral. In
this case it is not possible to set tenant links to the application
because we set as tenant links the tenant links of the components.
- Added check if application does not have components during the
transformation of the data.

Change-Id: I7bf1d3e1f96c7d995ead98aa777618f7262ef914
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/29707
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>