Apply filter on registries list when x-project header is passed
Related to VBV-1900

As part of the introduction of project-specific registries, we need to
have a way to get the registries that are available in the scope of a
given project. To achieve that, now a custom RegistryFactoryService is
added. When handling a GET request, it will check for the x-project
header. If it is not set, it will list all of the available registries
(some of the registries might not be visible to the current user due to
authz restrictions). If the x-project header is set, then the results
will be filtered so they will include only global registries and
registries that are specific to this project. Whether a registry is
global or not is determined by its tenantLinks. If there is a project
tenantLink present, the registry is specific to that project.
Otherwise, it is a global registry.

Change-Id: I74ede14e914de59fec4b39d07abfde446883469c
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/29413
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Georgi Muleshkov <gmuleshkov@vmware.com>