Allow projects to be deleted in standalone Admiral with no auth
Currently, when trying to delete a project, a check is done whether the
current user has the privilege to carry out the delete command. However,
when authentication is not enabled in Admiral, the security context
check fails and the delete call hangs.

Now, if the authentication is not enabled, no security context check
will be performed. Additionally, if the check is performed but fails,
the delete request will be failed as well (and will not be left
hanging).

Additionally, this commit includes an attempt to untangle some of the
branching in the ProjectService.handleDelete in order to make the code
more readable in the future.

Change-Id: I5d00ce51b77567f899032f7f474c664f4c7d33ce
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/30016
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Georgi Muleshkov <gmuleshkov@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>