[VBV-1905] Take into account registry's scope when searching for images
With the introduction of project-specific registries, we need to be able
to filter container image search results to images only from registries
that are defined as global or in the scope of the currently selected
project.

Now, GET requests to the TemplateSearchService (/templates endpoint) and
to ContainerImageTagsService (/images/tags endpoint) can optionally
include the x-project header. If this header is set, search results will
include only images that are accessible in the scope of the denoted
project (i.e. the are project-specific registries for this project or
are defined as global registries)

Change-Id: Ib0e87b85163549bba6b13d8384a801203e1f6032
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/30017
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Georgi Muleshkov <gmuleshkov@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>