[VBV-1954][Registries] Registry filter by path is not working well with https registries
The reason for the failure was the result of the
URI.getSchemeSpecificPart(). This works as expected when the URI
contains a schema. For example, new URI("http://example.com:80").getSchemeSpecificPart()
yields //example.com:80. However, in the case of a https registry, we
store the registry address without the schema part (because https is the
default schema). Thus, we end in the following case:
new URI("example.com:80").getSchemeSpecificPart() = 80. 80 is not the
right answer here and since it is used to filter the registry search
results, all results from this secure registry are truncated.

Change-Id: I498655de41489fbd3ffc2ef97827fdcec0f5b486
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/31217
Reviewed-by: Antonio Filipov <afilipov@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>