[VBV-1960] Disallow global registries to be deleted from project registries view
Currently, it is possible to select any registry in the project
registries view (regardless of its scope) and issue a delete command for
it. If you are the cloud admin, this request will succeed, if not, it
will fail with 403 and the red invalid session header will pop out.
Global registries are meant to be managed by cloud admins only and
through the Global registries view only. They are displayed in the
project registries view for info purposes only.

With this commit, it is now not possible to issue delete requests for global
registries from the project registries view. If the delete button is
pressed with at least one global registry in the current selection, a
warning message will be displayed to clarify that deletion of these
registries was skipped. Also, the delete confirmation popup now notes
that global registries cannot be deleted from this view.

VIDEO PREVIEW: http://go/5j7f

Change-Id: Ifa9b8bf8735fa1cde7a79d622eabd7f9c0305acc
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/31468
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Iveta Ilieva <iilieva@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>