Allow k8s clusters to be displayed in clusters view if not in PKS mode
If not in PKS mode with k8s support enabled, it is possible to add a k8s
cluster via the clusters view. However, it is not possible to see or
manage it after that because k8s clusters are filtered in this view.

This commit will allow the k8s clusters to be displayed in the clusters
view if we are not in PKS mode.

Change-Id: I3cc32b2e701815d45958d61b62bf1648c8f6da9f
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/32819
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Miroslav Shipkovenski <mshipkovenski@vmware.com>