VBV-1993 - [UPGRADE] After upgrade cannot add a host to a pre-upgrade created cluster (placement zone)
The problem was  in the transformation service during the upgrade. On
upgrade it is possible to change a resource pool to "ELASTIC" in this
case a query is executed in order to get the hosts and this query has a
clause to search computes with type "VM_GUEST". This is set for the
upgraded compute states but when creating new ones this type is not set
and not needed. This is why the new hosts are not visible in the UI.
Also when updating old hosts inside the cluster they will disappear from
the UI because the type will be removed from the compute state.
- Changed the EPZ not to set the type clause
- Changed the transformation not to set type for the compute states
- Added test for the scenario

Change-Id: I5e3f12be7d4e499b5e4e17144ca13e38565501f7
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/33218
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>