VBV-1993 - [UPGRADE] After upgrade cannot add a host to a pre-upgrade created cluster (placement zone) part 2
The problem was in the transformation service during the upgrade. On
upgrade it is possible to change a resource pool to "ELASTIC" in this
case a query is executed in order to get the hosts and this query has a
clause to search computes with type "VM_GUEST". This is set for the
upgraded compute states but when creating new ones this type is not set
and not needed. This is why the new hosts are not visible in the UI.
Also when updating old hosts inside the cluster they will disappear from
the UI because the type will be removed from the compute state.

The change should fix the issue where the upgrade already passed and the
newly added hosts to old clusters are not visible:
- Added bootstrap service that on start will check all the pools and for
the elastic ones type clause (VM_GUEST) will be removed from the query

Change-Id: I8426bd7322ebe78d4e0a67e389041dd88768afe1
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/33337
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>