Bump xenon to latest and add postgres
Changes in xenon that had to be addressed in the code/tests:
- The old migration was introduced again to the xenon framework but with
another endpoint (-old). Modified our code to continue using the old
migration since the new one does not work for older xenon versions
- Patch returns 304 not modified.
- OData access to FactoryService now always requires expand parameter to
expand documents. Previously, when GET request has any odata
parameter(e.g.: $count), response was always expanded regardless of
presence of expand parameter. It is changed to always
require $expand parameter in order to expand the response. This only
applies to GET on factory.
- Breaking change: When 304 is specified as status code, response
operation body becomes always null.
- ‘/’ in the document self links -> CounterSubtaskService kubernetis pod
log service

Fixed problems/tests with postgres:
- KubernetesAdapterServiceTest -> postgres is faster than lucene. Test
was waiting for availability of the bootstrap service but the service is
stopped before the check. -> Removed the check.
- DockerHostAdapterServiceTest -> “cannot determine table name” The
problem was that AbstractCallbackServiceHandler has persistence but it
is abstract and there are 3 services extending it. All the services does
not have factories so not tables created. Removed the persistence since
it is not needed. It is used for the data collection even in case a host
is restarted it is not a problem because new data collection will be
triggered. The states are not important since the task is auto deleted
after the task is completed.
- CompositeDescriptionServiceTest.testGetCompositeDescriptionWithDescriptionLinks
->
assertEquals(2,
retrievedCompositeDescriptionImages.descriptionImages.size()); ->
Postgres is faster than lucene. System container description is created
bore the check is executed.
- ContainerServiceTest.testVersionRetentionWtihOData -> There are no
different document versions with postgres -> Ignored the test
- ExposedPortsHostFilterTest.java - java.lang.AssertionError: Unexpected
exception: java.lang.IllegalStateException: Filtered hostLinks size is:
3 - Expected size is: 1 - The problem was that with posters query for
“ports.items.containerPort” does not work. Added filtering when filling
the results.
- RequestBrokerServiceTest.testContainerLoadBalancerLifeCycleMissingBackendContainer
-  org.junit.ComparisonFailure: expected:<[Allocation Filter Error: No
container descriptions with links [[wp]].]> but was:<[One of term,
booleanClauses must be provided]> -> createResourcesQuery method in
ContainerRemoveal task service was creating a query with empty list
values. It looks like it works with lucent but with postgres the query
fails
- ContainerRemovalTaskServiceTest - java.lang.AssertionError: Provided
description should be null! -> randomly failing - timing issue
reproducible also with lucene . When multiple containers with the same
description are deleted it could happen that the container description
is not deleted. Scheduled another attempt to delete it.
- SubscriptionManagementTest.java - java.lang.IllegalArgumentException:
Cannot determine SQL table name for document: /service/1951 - The test
was using MinimalTestState from core xenon. However for this state there
is no table in the database. Changed the test state to be container
state
- KubernetesEntityDataCollectionTest - java.lang.AssertionError:
expected:<1> but was:<0> -> there are also randomly failing tests - The
problem was the query was not created properly.

Added initial liquibase schema

Change-Id: I923296f5cf157c5277d8d929b2971b47816aa3a1
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/35916
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
PG-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>